requirements.txt
matplotlib==3.7.0
numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
Pillow==10.3.0
scikit_learn==1.2.1
scipy==1.8.0
seaborn==0.13.2
